Basic latent interaction framework
There's now a time value on the interaction struct. If this is not 0, when the interaction is triggered, the pawn can no longer interact for that time, giving the interaction component time to play an animation. (However, currently, the interaction component must call EndLatentInteraction() itself for this to happen. Perhaps in future a timer will be setup by the default impl that calls that automatically).
Also added some new utility functions and MeasurementAsString now uses SanitizeFloat instead of just default printf behaviour, meaning there's no longer a billion trailing zeros.