Refactor held item type to new base class
AResourceItem and AResourceContainer are now subclasses of AChemicodeObject, a base class with only a static mesh and outline component. This new base class allows the player told hold either a container or an item (and anymore objects that could be added in the future). Interactions with items and tooltips are acheived by attempting to cast the HeldItem or HighlightedItem to either ResourceItem or ResourceContainer and then doing the appropriate action depending on a null check. This is only done on tooltip update, which is only done when something else is hovered; no performance concerns.